requirements.txt
pytorch
numpy
scipy
argparse